fix: fix type in rest form dtos

+++ backend/src/main/java/format/backend/form/dto/FormDetailResponseDto.java
@@ -42,7 +42,7 @@ public record FormDetailResponseDto(
 
         @NonNull List<@NonNull QuestionResponseDto> questions,
 
-        @NonNull Integer ratingCount,
+        @NonNull Long ratingCount,
 
         @NonNull Double ratingAvg,
